requirements.txt
streamlit==1.15.0
opencv-python==4.5.5.64
numpy==1.24.2
scikit-image==0.19.3
Pillow==9.3.0
scikit-learn==1.1.2
tqdm==4.64.1
matplotlib==3.5.1
pandas==1.5.3